# web/openapi.json
{
  "components": {
    "schemas": {
      "Approval": {
        "description": "One approval envelope: subject, content hash, verdict, expiry.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ApprovalKind"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "payload_hash": {
            "title": "Payload Hash",
            "type": "string"
          },
          "policy": {
            "title": "Policy",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApprovalStatus"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "kind",
          "subject",
          "payload",
          "payload_hash",
          "status",
          "policy",
          "reason",
          "decided_by",
          "created_at",
          "expires_at",
          "decided_at"
        ],
        "title": "Approval",
        "type": "object"
      },
      "ApprovalKind": {
        "description": "What is being approved. ``PLAN`` gates the task; the rest gate one tool call.",
        "enum": [
          "plan",
          "shell",
          "push"
        ],
        "title": "ApprovalKind",
        "type": "string"
      },
      "ApprovalStatus": {
        "enum": [
          "pending",
          "approved",
          "rejected",
          "expired"
        ],
        "title": "ApprovalStatus",
        "type": "string"
      },
      "Decision": {
        "enum": [
          "approve",
          "reject"
        ],
        "title": "Decision",
        "type": "string"
      },
      "DecisionRequest": {
        "description": "A human's answer. ``payload_hash`` is optional but verified when sent.",
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "payload_hash": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Hash"
          }
        },
        "required": [
          "decision"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "Event": {
        "description": "One append-only audit record. ``id`` is monotonic and doubles as the SSE id.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "type",
          "payload",
          "ts"
        ],
        "title": "Event",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Task": {
        "properties": {
          "cost_usd": {
            "default": 0.0,
            "title": "Cost Usd",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "issue_url": {
            "title": "Issue Url",
            "type": "string"
          },
          "repo_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Sha"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "steps": {
            "default": 0,
            "title": "Steps",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "issue_url",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Task",
        "type": "object"
      },
      "TaskCreate": {
        "properties": {
          "issue_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Issue Url",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "issue_url"
        ],
        "title": "TaskCreate",
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "queued",
          "planning",
          "awaiting_approval",
          "solving",
          "verifying",
          "delivering",
          "done",
          "failed",
          "rejected"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "RepoAegis",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/approvals/{approval_id}": {
      "get": {
        "operationId": "get_approval_api_approvals__approval_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approval"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Approval"
      }
    },
    "/api/approvals/{approval_id}/decision": {
      "post": {
        "description": "Answer one gate.\n\nRepeating the same answer is a no-op that returns the same envelope, so a\ndouble-click or a client retry cannot advance the task twice. A different\nanswer to an already-decided gate is a conflict, not an overwrite.",
        "operationId": "decide_approval_api_approvals__approval_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "default": "anonymous",
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approval"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Approval"
      }
    },
    "/api/events": {
      "get": {
        "operationId": "stream_events_api_events_get",
        "parameters": [
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Last-Event-Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events"
      }
    },
    "/api/healthz": {
      "get": {
        "operationId": "healthz_api_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthz Api Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz"
      }
    },
    "/api/tasks": {
      "get": {
        "operationId": "list_tasks_api_tasks_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "title": "Response List Tasks Api Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Tasks"
      },
      "post": {
        "operationId": "create_task_api_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task"
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task"
      }
    },
    "/api/tasks/{task_id}/approvals": {
      "get": {
        "operationId": "list_task_approvals_api_tasks__task_id__approvals_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Approval"
                  },
                  "title": "Response List Task Approvals Api Tasks  Task Id  Approvals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Task Approvals"
      }
    },
    "/api/tasks/{task_id}/events": {
      "get": {
        "operationId": "list_task_events_api_tasks__task_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  },
                  "title": "Response List Task Events Api Tasks  Task Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Task Events"
      }
    }
  }
}